requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
dataproperty==1.1.0
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jieba3k==0.35.1
jiter==0.9.0
joblib==1.4.2
lxml==5.4.0
lxml-html-clean==0.4.2
markdown-it-py==3.0.0
mbstrdecoder==1.1.4
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.9.1
openai==1.76.2
packaging==25.0
pathvalidate==3.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
setuptools==80.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
structlog==25.3.0
tabledata==1.3.4
tavily-python==0.7.1
tcolorpy==0.1.7
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typepy==1.3.4
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
